+++ application/src/main/java/org/thingsboard/server/service/entitiy/cf/CalculatedFieldCtxId.java
@@ -15,17 +15,7 @@
  */
 package org.thingsboard.server.service.entitiy.cf;
 
-import org.thingsboard.server.common.data.cf.CalculatedField;
-import org.thingsboard.server.common.data.exception.ThingsboardException;
-import org.thingsboard.server.common.data.id.CalculatedFieldId;
-import org.thingsboard.server.service.security.model.SecurityUser;
+import java.util.UUID;
 
-public interface TbCalculatedFieldService {
-
-    CalculatedField save(CalculatedField calculatedField, SecurityUser user) throws ThingsboardException;
-
-    CalculatedField findById(CalculatedFieldId calculatedFieldId, SecurityUser user);
-
-    void delete(CalculatedField calculatedField, SecurityUser user);
-
+public record CalculatedFieldCtxId(UUID cfId, UUID entityId) {
 }
