+++ common/dao-api/src/main/java/org/thingsboard/server/dao/cf/CalculatedFieldService.java
@@ -60,4 +60,6 @@ public interface CalculatedFieldService extends EntityDaoService {
 
     boolean referencedInAnyCalculatedField(TenantId tenantId, EntityId referencedEntityId);
 
+    boolean existsCalculatedFieldByEntityId(TenantId tenantId, EntityId entityId);
+
 }
