+++ dao/src/main/java/org/thingsboard/server/dao/cf/CalculatedFieldDao.java
@@ -34,4 +34,6 @@ public interface CalculatedFieldDao extends Dao<CalculatedField> {
 
     List<CalculatedField> removeAllByEntityId(TenantId tenantId, EntityId entityId);
 
+    boolean existsByEntityId(TenantId tenantId, EntityId entityId);
+
 }
